requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.21.0
scipy>=1.7.0
einops>=0.6.0

# Configuration and logging
hydra-core>=1.3.0
pyyaml>=6.0
wandb>=0.15.0
tqdm>=4.65.0
matplotlib>=3.5.0

# Data processing and metrics
scikit-learn>=1.0.0
pandas>=1.3.0

# Development and testing (optional)
pytest>=7.0.0
black>=22.0.0
isort>=5.10.0
mypy>=0.910
flake8>=4.0.0

# Documentation (optional)
sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0 